+++ R/make_graph_login.R
@@ -1,4 +1,4 @@
-make_graph_login_from_token <- function(token, aad_host, graph_host)
+make_graph_login_from_token <- function(token, azure_host, graph_host)
 {
     if(is_empty(graph_host))
         return()
@@ -7,7 +7,7 @@ make_graph_login_from_token <- function(token, aad_host, graph_host)
     newtoken <- token$clone()
     if(is_azure_v1_token(newtoken))
         newtoken$resource <- graph_host
-    else newtoken$scope <- sub(aad_host, graph_host, newtoken$scope, fixed=TRUE)
+    else newtoken$scope <- c(paste0(graph_host, ".default"), "openid", "offline_access")
 
     newtoken$refresh()