+++ tests/testthat/test01_auth.R
@@ -9,7 +9,10 @@ if(tenant == "" || app == "" || password == "" || subscription == "")
     skip("Authentication tests skipped: ARM credentials not set")
 
 
+clean_token_directory(confirm=FALSE)
 suppressWarnings(file.remove(file.path(AzureR_dir(), "arm_logins.json")))
+
+scopes <- c("https://management.azure.com/.default", "openid", "offline_access")
 
 test_that("ARM authentication works",
 {
@@ -17,7 +20,7 @@ test_that("ARM authentication works",
     expect_is(az, "az_rm")
     expect_true(is_azure_token(az$token))
 
-    tok <- get_azure_token("https://management.azure.com/", tenant, app, password)
+    tok <- get_azure_token(scopes, tenant, app, password, version=2)
     az2 <- az_rm$new(token=tok)
     expect_is(az2, "az_rm")
 })
@@ -40,7 +43,7 @@ test_that("Login interface works",
     az5 <- get_azure_login(tenant)
     expect_is(az5, "az_rm")
 
-    tok <- get_azure_token("https://management.azure.com/", tenant, app, password)
+    tok <- get_azure_token(scopes, tenant, app, password, version=2)
     az6 <- create_azure_login(token=tok, graph_host=NULL)
     expect_is(az6, "az_rm")
 })
@@ -58,6 +61,10 @@ test_that("Graph interop works",
 
     gr <- AzureGraph::get_graph_login(tenant)
     expect_is(gr, "ms_graph")
+    expect_true(
+        (!is.null(gr$token$resource) && grepl("graph\\.microsoft\\.com", gr$token$resource)) ||
+        (!is.null(gr$token$scope) && any(grepl("graph\\.microsoft\\.com", gr$token$scope)))
+    )
 })
 
 test_that("Top-level do_operation works",